Clean-up samples admin view
